removed pyenv calls in versioner and packages
These don't work because the environment can't be passed by variable.
But, since these are executed by the head node calling ducttape, a simple
workaround is to make sure the python environment is loaded when ducttape
is run. This avoids needing to redundantly set the environment for every
package and versioner, which isn't intended to handle that capability.
Instead, the packages will be installed into the calling environment,
and all tasks will be executed in that environment so long as $pyenv
is set in the tconf.