major refactoring
ducttape is great, but it's hard to manage complex workflow logic.
therefore, this refactoring puts the logic inside the steps, rather
that outside. all steps take in both source and target. as much
as possible, steps are combined, and logic is put within the shell
script. dev is handled the same as train. test also gets processed,
but test decoding uses decode_with_bundle, so that the separate
processing that typically gets applied to test data can be handled
there.